Add Plumbing to attempt to reopen the PipeWire picker
Adds a "FocusList" function to the DesktopCapturer interface that can
be used to notify capturers that they are now in focus. This can be used
(as it is here) for capturers like the PipeWire picker to see if they've
previously been dismissed and should attempt to re-create themselves.

In order to properly re-create itself the PipeWire picker needs to do a
couple of things:
1) Stop any existing (previous) stream choice
2) Increment to a new source id, so that any existing restore token is
   overwritten.

In order to properly stop the previous stream choice, the shared
screencast stream needed to be modified to clear the existing queue upon
receiving a Stop event, and it further ensures that we only send blank
frames out if we do not have an active PipeWire connection.

Bug: chromium:1351577
Change-Id: Iacaf6311a46e5751792c16fa909c44821f259efd