dcsctp: Add handover state for stream counts
To allow the transport to be able to know which ranges of
stream identifiers it can be use, the negotiated incoming/inbound
and outgoing/outbound stream counts will be exposed. They are
added to Metrics, and guaranteed to be available from within
the OnConnected callback.

Note that dcSCTP will not validate that the client is sending
on a stream that is within the negotiated bounds. Sending can
be done before the association has been established.

Bug: webrtc:14277
Change-Id: Ic764e5f93f53d55633ee547df86246022f4097cf