Avoid duplicate RTCCodecStats entries.
The code incorrectly assumed that codecs exist on a per-mid/transceiver
basis, but codec payload types are unique on a per-transport basis and
in practise most applications use BUNDLE (single transport for the
entire PC).

This CL makes the codecs per-transport instead of per-transceiver. We
still iterate transceivers because codecs are exposed on a
per-transceiver basis, but using std::set<PayloadType> we can
efficiently filter out duplicates without hundreds of string
comparisons.

Manual testing shows that this CL brings down the number of "codec"
stats in Google Meet 50p from 872 objects to 43 objects.

Bug: webrtc:14414
Change-Id: Ic854b31bd595799554b99fff22cbd48264ebd141