dcsctp: Specify an initial RTT
The RTT will quickly be updated whenever a DATA chunk is acked, but
the initial value was set to zero. In unit tests, which often are
short and rarely increase the simulated time between a DATA is sent,
and acked, the smoothed RTT would usually stay at 0, which causes
e.g. the rate limiting of FORWARD not to work.

Bug: None
Change-Id: Ieb515fe875ce88d001777b00d6efd9762565a09d