Add test coverage for P2PTransportChannel refactor (#8/n)
This converts all P2PTransportChannel unit tests to parameterized tests, with a string parameter for the field_trials which is used to enable the refactor. This adds a variation of each existing test using the refactored code path.

Some tests depend on the order of connections returned from P2PTransportChannel::connections(), which earlier mirrored the internal connections order of the ICE controller. These tests have been rewritten to not depend on this internal detail.

Bug: webrtc:14367, webrtc:14131
Change-Id: I0469550d571ed389804eb486fe5bd22504e59373