Fix TimestampScaler for sample rate < clock rate
Two cases:
1. When using a codec where info.SampleRateHz() < info.GetFormat().clockrate_hz,
when the external codec wraps around, the scaled timestamp incorrectly jumps.
example:
 external: 4294966297	scaled: 4294964743   expected: 4294964743
 external: 4294966630	scaled: 4294964854   expected: 4294964854
 external: 4294966963	scaled: 4294964965   expected: 4294964965
 external: 0         	scaled: 2863309311   expected: 4294965076, error
 external: 333       	scaled: 2863309422   expected: 4294965187, error
 external: 666       	scaled: 2863309533   expected: 4294965298, error

This corrects for that. None of the existing codec options have this property, but we noticed this issue when using a custom one.

2. When using a codec with scaling like G722, ToExternal is incorrect
when the internal timestamp wraps around. This was not noticed in
the tests because ToInternal was always called before ToExternal.
Because ToInternal is stateful, it made it so internal_timestamp was always the same as internal_ref_ inside of ToExternal, never actually
testing the case where there is a diff. The G722 unittests failed when calling ToInternal first

Bug:None
Change-Id: I4f151406ff9d27decdb5acfeb00bca7d177d911f