SharedScreenCastStream test: increase waiting times
This doesn't effect for how long the test will run, it just gives
PipeWire more time to establish connection and create empty buffers
before we try to work with it. All the waiting events will be
interrupted by signals once we no longer need to wait so it doesn't
matter if we wait 2 seconds or 5 seconds.

Bug: webrtc:14568
Change-Id: Ie918e8943bf882059b1289f57595fc302216745e