AudioProcessingImpl: Add the use of AGC2 InputVolumeController
The integration relies on GainController2 methods Analyze(), Process(),
and GetRecommendedInputVolume() to internally take into account whether
the input volume controller is enabled in the ctor or not. These methods
are called for every frame processed if GainController2 is enabled.

The functionality can be enabled from the APM config and is not enabled
by default. If both AGC1 analog controller (gain_controller1 and
analog_gain_controller) and AGC2 input volume controller
(gain_controller2 and input_volume_controller) are enabled from the
config, AGC2 input volume controller gets a higher priority and
AgcManagerDirect methods AnalyzePreProcess() and Process() are not
called during processing.

Bug: webrtc:7494
Change-Id: I9105483be34eb95fab3c46afbbd368802e956fad