AGC2 `InputVolumeController`: tests to repro a potential issue
The tests added in this CL should pass; however, the first step of
each test attempts to trigger a downward input volume adjustment
caused by clipping input. The only effect is that the maximum input
volume is lowered, but the recommended volume is not lowered until
speech is observed.

TODO: replicate the test for `AgcManagerDirect` and, if the issue
is not there, fix `InputVolumeController`.

Bug: webrtc:14761
Change-Id: I1795e74c5f219e15107e928ebaca2bfa75214526