AGC1 and AGC2 input volume controllers do not step back on clipping
The tests added in this CL include a part that triggers a downward
input volume adjustment caused by clipping input. Both AGC1 and AGC2
controllers react on that by lowering the maximum input volume, but
the recommended input volume is not lowered.

If the behavior is intended, abandon this change; otherwise fix the
behavior in `InputVolumeController` and remove the tests for
`AgcManagerDirect`.

Bug: webrtc:14761
Change-Id: I1795e74c5f219e15107e928ebaca2bfa75214526