Add integration test for NACK functionality
This adds a test that sets the required feedback mechanisms
to get NACK configured for video, connects, and then sets packet
loss to 100%.

The expected result is that the receiver sends NACK; this will cause
the test to set packet loss to 0%, so the next NACK sent should get
to the sender and cause retransmission.

This is explicating a problematic case in splitting media channel.

Bug: webrtc:13931
Change-Id: I0c23c4a89953976454d84b0211f0a7545bbb717a