De-flake NonSenderRttStats and make it faster to run.
It takes several seconds until we get an RTT measurement because that
requires RTCP packets to be received and those are not sent very often.

With a timeout of 10s, this passes most of the time but occasionally
fails on bots. This CL gives the test 20s, but to avoid this test
taking 20s to run every time the update the test to stop blocking as
soon as an RTT measurement is received (which usually takes around 5s).

Bug: webrtc:14981
Change-Id: Ieca94c90dfb52c3b17584a06660ff66c6462aa8b