PipeWire capturer: download pw_buffer screen data once requested
We currently process all buffers as they go and keep just the most
recent frame. In case the compositor is faster than the browser it can
happen we process buffers that are not going to be used at all, which is
just waste of processing power. Keep the latest unprocessed buffer and
process it until it's really needed to save some performance.

Submitting on behalf of Martin Stránský.

Bug: webrtc:13429
Change-Id: Ibb7f4c3e7e1968409a4fa9a151ad4b13257b4cb5