PipeWire capturer: import DMABufs directly into desktop frame
Originally DMABufs were imported into a temporary buffer followed by a
copy operation into the desktop frame itself. This is not needed as we
can import them directly into desktop frames and avoid this overhead.

Also drop support for MemPtr buffers as both Mutter and KWin don't seem
to support them and they are going to be too slow anyway.

Bug: webrtc:13429


PipeWire capturer: download pw_buffer screen data once requested

We currently process all buffers as they go and keep just the most
recent frame. In case the compositor is faster than the browser it can
happen we process buffers that are not going to be used at all, which is
just waste of processing power. Keep the latest unprocessed buffer and
process it until it's really needed to save some performance.

Submitted on behalf of Martin Stránský.

Bug: webrtc:13429
Change-Id: I4242b6be87ea06aca8ccb42e915a152cd22ef68a